Updated structure to hold all the cards in a package
